+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 `fixnum` adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Removed
+- Support of `fixnum::ops::Numeric` (@quasiyoke).
+
+### Added
+- Traits `One`, `Zero`, `Bounded`.
+- Saturating operations: `saturating_add`, `saturating_sub`, `saturating_rmul` (@quasiyoke).
+- Docs for `FixedPoint::integral` method (@quasiyoke).
+
+## 0.2.3 - 2020-12-30
 ### Added
 - `Compact` implementation for `parity` feature (@quasiyoke).
 - `Clone` implementation for errors (@quasiyoke).